Switch to opensearch
Part of the preparation to migrate off elasticsearch. GOV.UK PaaS will drop support for elasticsearch soon. See https://docs.cloud.service.gov.uk/deploying_services/opensearch/#migrating-from-elasticsearch.

We need to disable security so that we can continue accessing it via HTTP. Otherwise, we'd need to start using HTTPS. I think it uses a self-signed certificate, so we'd need to make a lot of changes to get that to work.