Add database schema dump
The database dump contains pseudonymised user data, so is only available to people within the Digital Marketplace team. Create a dump of the schema that contains no user data. This is safe for us to make public and share with the world. Other people can use this instead of the cleaned database dump if they can't access it.

Update the setup code to point people to the new file if they can't access the main dump.

To review, use `gunzip` to convert the file to sql.